Configurable make, preserve makeprg, key map.
The tdd plugin is more configurable; an errorformat must be added and
Tdd_makeprg must be set, see the inline doc on 'usage'.

Original makeprg setting is preserved, Tdd_makeprg is only set when test
needs to be run.

In vimrc the key map for task list was changed to ,l and ,t key map was
added to run test.