KISS approach for setting the runner.
At first I was thinking about a system that would be able to switch
between its test file and the file of the component under test, but that
would be over the top, too much effort would be required to build it.
Instead, just manipulating the g:Tdd_makeprg variable to set it to a
certain test run command, is far more simple and elegant solution to the
problem of running a test while editing another file.