Revert wrong assumption.
The assumption that less eye movement was needed to track the cursor was
wrong.  Maybe not in the sense of pure eye movement, but shifting the
viewport instead of the cursor through the document works confusing.
Especially when you are jumping between the only 2 occurences in a file
that are only a few lines away from each other.