Update script change.
Since 1.7.8 git handles submodule differently, see
https://raw.github.com/gitster/git/master/Documentation/RelNotes/1.7.8.txt
search for 'When populating a new submodule directory'.

This made the find for directories named '.git' fail in the update
script for newly added submodules...
We could broaden the find command into finding directories and files,
but using git's own system feels like a more natural choice.

This will not take into account plugins in bundle/ that we are just
testing (ie. are not yet under versioning control), but since it's just
for testing, I don't mind that they're not.