A more useful way of line numbering.
I've always had mixed feelings about line numbering gutters; When you
want to know at which line you are, the status line can tell you (or a
key combination like <c-g>.  What I have found to come in handy from
time to time is having to go/down a certain amount of lines, or having
an indicator of the number to use with a motion command.
Since 7.3 there is apparently relativenumber and I have a feeling it
might come in handy, so let's try it.