And back to the original ,b it is...
On top of the reasons listed in commit
4031a4346c5a5426aa5a41852ddab10751360d41 the biggest problem, IMO, with
using ctrlp as a buffer selector is that you cannot have a split-window
view on the same buffer; imagine the following:

A vertical split with on the left a.txt and on the right b.txt, the
window with a.txt is the current one.
You want the b.txt buffer in the current window, so you select b.txt
with ctrlp (which is possible since a.txt is loaded) and press enter;
What happens is that you are switched to the other window, with b.txt
in, which is undesirable behaviour AFAIC; I want b.txt loaded in the
current window.