Airline: show icon only on trailing whitespace.
The algorithms of Airline are suboptimal at the moment to do decent
enough mixed indent checking IMO; either the error was showing almost
constantly (confusing TAB indent with spaces in comments for alignment
(or doc comment blocks) or it showed no error message because indenting
whitespace wasn't mixed on 1 line... For instance: adding a line with
spaces-only indentation in a file that has TAB indentation, did not show
the error, which is why I would want this capability to begin with.
Therefor => disabling it for the time being.

Furthermore, since now the whitespace error is only for trailing
whitespace, the message itself can go, the icon on itself will tell me
enough, and I save some valuable status line space.