Move git commit spell checking activation from ftplugin to autocmd.
Setting it in a ftplugin/ file did the trick for spelling, but it fucked
up text wrapping (it overrides I guess).  Working with an autocmd just
adds spell checking to the commit message buffer, but preserves the
automatic text wrapping.
Could probably figure this out and fix it in a ftplugin/ file, but don't
have the time at the moment.