Remove airline plugin.
While Airline is a nice plugin, I'm realising that all that info in the
status all of the time just isn't necessary; some basic info like file
path, file type and position in the buffer are enough when editing.

However; it _is_ necessary to be able to request more information on the
current buffer/file/directory, this is where the functions#buffer_info()
together with the :Info command come into play (see [4c6f3dd]).